fix ci

+++ apg_tga/test_code/main_write.c
@@ -16,14 +16,14 @@ int main( int argc, char** argv ) {
     return 0;
   }
   printf( "loading %s\n", argv[1] );
-  unsigned int w, h, n;
+  int w, h, n;
   uint8_t* img = stbi_load( argv[1], &w, &h, &n, 0 );
   if ( !img ) { return 1; }
 
   printf( "loaded %s - %ix%i n=%i!\n", argv[1], w, h, n );
-  if(!apg_tga_bgr_to_rgb( img, w,h,n)){return 1;}
+  if(!apg_tga_bgr_to_rgb( img, (unsigned int)w,(unsigned int)h,(unsigned int)n)){return 1;}
 
-  if ( !apg_tga_write_file( "out.tga", img, w, h, n ) ) {
+  if ( !apg_tga_write_file( "out.tga", img, (unsigned int)w, (unsigned int)h, (unsigned int)n ) ) {
     fprintf( stderr, "ERROR writing file\n" );
     return 1;
   }
